完成运动、心率显示功能

+++ SMA/SMASedentEditCell.m
@@ -222,7 +222,7 @@
         _switchBlock(sender,_alarmInfo);
     }
     else{
-       _alarmSwitch.on = _alarmInfo.isOpen;
+       _alarmSwitch.on = _alarmInfo.isOpen.intValue;
     }
 }
